requirements.txt
# ============================
# Core API
# ============================
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# ============================
# Database
# ============================
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.1

# ============================
# Validation & Settings
# ============================
pydantic>=2.6.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1

# ============================
# Security
# ============================
python-jose[cryptography]>=3.3.0
passlib==1.7.4
bcrypt==3.2.2


# ============================
# Utilities
# ============================
email-validator>=2.1.0
jinja2>=3.1.2
python-dotenv



# ============================
# Testing
# ============================
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0